fixed glVertexAttribute on shader side and enhanced test() func
Also, isoValue is now a vec3 to render 3 contours.

+++ GuidedResearch/vert.glsl
@@ -1,6 +1,6 @@
 #version 420 core
 
-in layout(location = 0) ivec2 scalarFieldCoords;
+in layout(location = 0) vec2 scalarFieldCoords;
 
 void main() {
   // Store coordinates in gl_Position.xy
